pyproject.toml
[project]
name = "audio-transcript"
version = "0.1.0"
description = "Multi-provider transcription API with Flask, Redis, Groq, Mistral, and whisper.cpp fallback."
requires-python = ">=3.10"
dependencies = [
  "Flask>=3.0.0",
  "psycopg[binary]>=3.2.0",
  "pyarrow>=18.0.0",
  "python-dotenv>=1.0.0",
  "redis>=5.0.0",
  "requests>=2.31.0",
  "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
